+++ app/Casts/Portforward.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace App\Casts;
+
+use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Log;
+use App\Models\Sys\Vm;
+
+class Portforward
+{
+    /**
+     * Cast the given value.
+     *
+     * @param  array<string, mixed>  $attributes
+     */
+    public function get(Vm $model, string $key, mixed $value, array $attributes)
+    {
+        foreach ($model->clients as $vm) {
+            $value .= "\n";
+            $value .= implode(':', [$vm->pri_host, implode(':', $vm->portopen)]);
+        }
+        $value = trim($value);
+        return $value;
+    }
+
+    /**
+     * Prepare the given value for storage.
+     *
+     * @param  array<string, mixed>  $attributes
+     */
+    public function set(Vm $model, string $key, mixed $value, array $attributes)
+    {
+        return $value;
+    }
+}